Moved add_images
