+++ deepomatic/cli/cmds/platform/utils.py
@@ -1,2 +1,3 @@
 from deepomatic.cli.lib.platform import PlatformManager  # noqa: F401
 from deepomatic.cli.lib.inference import InferManager, DrawImagePostprocessing, BlurImagePostprocessing  # noqa: F401
+from deepomatic.cli.lib.add_images import AddImageManager  # noqa: F401
